+++ NEWS.md
@@ -1,3 +1,12 @@
+# multid 0.5.0
+
+* Added option to obtain scaled estimates in ml_dadas. Scaling is done for both difference score components and the difference scores, based on random intercept SDs and random slope SD, respectively, in a reduced model without the predictor and the interaction between predictor and moderator
+* Added option to test for random effect covariation with likelihood ratio test in ml_dadas from a reduced model without the predictor and the interaction between predictor and moderator
+* Added option to include covariates in sem_dadas
+* Added variance test with sem in sem_dadas
+* Added variance test via parametric bootstrap in ml_dadas
+* Added cvv_manual -function for calculation of coefficients of variance variation from manually inputted sample sizes and variances of multiple variables
+
 # multid 0.4.0
 
 * Replaced two-sided tests in sem_dadas for absolute parameters with one-sided tests